feat: add recipes available functionality

+++ src/main/java/com/charapadev/secondchef/dtos/ShowItemDTO.java
@@ -7,6 +7,6 @@ public record ShowItemDTO(
 
     String name,
 
-    int quantity
+    long quantity
 ) {
 }
